Re order helper methods and fixtures
