+++ tests/integration/test_enrollment/conftest.py
@@ -2,12 +2,15 @@ import os
 import platform
 import socket
 import yaml
+import pytest
+import subprocess
+import ssl
 
 from wazuh_testing.tools import LOG_FILE_PATH, WAZUH_PATH
-import subprocess
 from wazuh_testing.tools.monitoring import FileMonitor
 from wazuh_testing.tools.configuration import load_wazuh_configurations
 from wazuh_testing.tools.configuration import set_section_wazuh_conf, write_wazuh_conf
+from wazuh_testing.tools.monitoring import ManInTheMiddle
 
 test_data_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'data')
 configurations_path = os.path.join(test_data_path, 'wazuh_conf.yaml')
@@ -193,3 +196,81 @@ def override_wazuh_conf(configuration, test):
     test_config = set_section_wazuh_conf(conf[0]['sections'])
     # Set new configuration
     write_wazuh_conf(test_config)
+
+tests = load_tests(os.path.join(test_data_path, 'wazuh_enrollment_tests.yaml'))
+
+LAST_MESSAGE = None
+CURRENT_TEST_CASE = {}
+
+def receiver_callback(received):
+    if len(received) == 0:
+        return b""
+
+    global LAST_MESSAGE
+    global CURRENT_TEST_CASE
+    LAST_MESSAGE = received.decode()
+    socket_listener.event.set()
+    response = CURRENT_TEST_CASE['message']['response'].format(**DEFAULT_VALUES).encode()
+    return response
+
+def get_last_message():
+    global LAST_MESSAGE
+    import time
+    timeout = time.time() + 20 # 20 seconds timeout
+    while not LAST_MESSAGE and time.time() <= timeout:
+        pass
+    return LAST_MESSAGE
+
+def clear_last_message():
+    global LAST_MESSAGE
+    LAST_MESSAGE = None
+
+socket_listener = ManInTheMiddle(address=(DEFAULT_VALUES['manager_address'], DEFAULT_VALUES['port']), family='AF_INET',
+                                              connection_protocol='SSL', func=receiver_callback)
+
+@pytest.fixture(scope="function")
+def configure_socket_listener():
+    socket_listener.start()
+    socket_listener.listener.set_ssl_configuration(connection_protocol=ssl.PROTOCOL_TLSv1_2,
+                                                   certificate='/var/ossec/etc/manager.cert',
+                                                   keyfile='/var/ossec/etc/manager.key',
+                                                   options=None,
+                                                   cert_reqs=ssl.CERT_OPTIONAL)
+
+    while not socket_listener.queue.empty():
+        socket_listener.queue.get_nowait()
+    socket_listener.event.clear()
+
+    yield
+    socket_listener.shutdown()
+
+@pytest.mark.parametrize('test_case', [case for case in tests])
+@pytest.fixture(scope="function")
+def set_keys(test_case):
+    keys = test_case.get('pre_existent_keys', [])
+    if not keys:
+        return
+    # Write keys
+    try:
+        with open(CLIENT_KEYS_PATH, "w") as keys_file:
+            for key in keys:
+                keys_file.writelines(key)
+    except IOError:
+        raise
+
+@pytest.mark.parametrize('test_case', [case for case in tests])
+@pytest.fixture(scope="function")
+def set_test_case(test_case):
+    global CURRENT_TEST_CASE
+    CURRENT_TEST_CASE = test_case
+
+@pytest.mark.parametrize('test_case', [case for case in tests])
+@pytest.fixture(scope="function")
+def set_pass(test_case):
+    # Write password file
+    try:
+        with open(AUTHDPASS_PATH, "w") as f:
+            if 'password_file_content' in test_case:
+                f.write(test_case['password_file_content'])
+    except IOError:
+        raise
