+++ tests/integration/test_enrollment/test_agentd_enrollment.py
@@ -12,99 +12,24 @@ from conftest import *
 # Marks
 pytestmark = [pytest.mark.linux, pytest.mark.win32, pytest.mark.tier(level=0), pytest.mark.agent]
 
-tests = load_tests(os.path.join(test_data_path, 'wazuh_enrollment_tests.yaml'))
 configurations = load_wazuh_configurations(configurations_path, __name__)
 
-LAST_MESSAGE = None
-CURRENT_TEST_CASE = {}
-
-def receiver_callback(received):
-    if len(received) == 0:
-        return b""
-
-    global LAST_MESSAGE
-    LAST_MESSAGE = received.decode()
-    socket_listener.event.set()
-    response = CURRENT_TEST_CASE['message']['response'].format(**DEFAULT_VALUES).encode()
-    return response
-
-def get_last_message():
-    global LAST_MESSAGE
-    import time
-    timeout = time.time() + 20 # 20 seconds timeout
-    while not LAST_MESSAGE and time.time() <= timeout:
-        pass
-    return LAST_MESSAGE
-
-def clear_last_message():
-    global LAST_MESSAGE
-    LAST_MESSAGE = None
-
-socket_listener = ManInTheMiddle(address=(DEFAULT_VALUES['manager_address'], DEFAULT_VALUES['port']), family='AF_INET',
-                                              connection_protocol='SSL', func=receiver_callback)
-
-# fixtures
+# Fixtures
 @pytest.fixture(scope="module", params=configurations)
 def get_configuration(request):
     """Get configurations from the module"""
     return request.param
 
-@pytest.fixture(scope="function")
-def configure_socket_listener():
-    socket_listener.start()
-    socket_listener.listener.set_ssl_configuration(connection_protocol=ssl.PROTOCOL_TLSv1_2,
-                                                   certificate='/var/ossec/etc/manager.cert',
-                                                   keyfile='/var/ossec/etc/manager.key',
-                                                   options=None,
-                                                   cert_reqs=ssl.CERT_OPTIONAL)
-
-    while not socket_listener.queue.empty():
-        socket_listener.queue.get_nowait()
-    socket_listener.event.clear()
-
-    yield
-    socket_listener.shutdown()
 
 @pytest.mark.parametrize('test_case', [case for case in tests])
-@pytest.fixture(scope="function")
-def set_keys(test_case):
-    keys = test_case.get('pre_existent_keys', [])
-    if not keys:
-        return
-    # Write keys
-    try:
-        with open(CLIENT_KEYS_PATH, "w") as keys_file:
-            for key in keys:
-                keys_file.writelines(key)
-    except IOError as exception:
-        raise
-
-@pytest.mark.parametrize('test_case', [case for case in tests])
-@pytest.fixture(scope="function")
-def set_pass(test_case):
-    # Write password file
-    try:
-        with open(AUTHDPASS_PATH, "w") as f:
-            if 'password_file_content' in test_case:
-                f.write(test_case['password_file_content'])
-    except IOError:
-        raise
-
-@pytest.mark.parametrize('test_case', [case for case in tests])
-@pytest.fixture(scope="function")
-def set_test_case(test_case):
-    global CURRENT_TEST_CASE
-    CURRENT_TEST_CASE = test_case
-
-@pytest.mark.parametrize('test_case', [case for case in tests])
-def test_agentd_enrollment(set_test_case, configure_socket_listener, configure_environment, set_keys, set_pass):
-    if 'wazuh-agentd' in CURRENT_TEST_CASE.get("skips", []):
+def test_agentd_enrollment(set_test_case, configure_socket_listener, configure_environment, set_keys, set_pass, test_case: list):
+    if 'wazuh-agentd' in test_case.get("skips", []):
         pytest.skip("This test does not apply to agentd")
 
     control_service('stop', daemon='wazuh-agentd')
-    override_wazuh_conf(CURRENT_TEST_CASE.get('configuration', {}), __name__)
+    override_wazuh_conf(test_case.get('configuration', {}), __name__)
 
-    if 'expected_error' in CURRENT_TEST_CASE:
+    if 'expected_error' in test_case:
         clean_log_file()
         try:
             control_service('start', daemon='wazuh-agentd')
@@ -112,7 +37,7 @@ def test_agentd_enrollment(set_test_case, configure_socket_listener, configure_e
             pass
         try:
             log_monitor = FileMonitor(LOG_FILE_PATH)
-            log_monitor.start(timeout=120, callback=lambda x: wait_until(x, CURRENT_TEST_CASE.get('expected_error')))
+            log_monitor.start(timeout=120, callback=lambda x: wait_until(x, test_case.get('expected_error')))
         except TimeoutError as err:
             assert False, f'Expected error log doesn´t occurred'
 
@@ -121,7 +46,7 @@ def test_agentd_enrollment(set_test_case, configure_socket_listener, configure_e
         control_service('start', daemon='wazuh-agentd')
         result = get_last_message()
         assert result != None, "Enrollment request message never arraived"
-        assert result == CURRENT_TEST_CASE['message']['expected'].format(**DEFAULT_VALUES),  \
+        assert result == test_case['message']['expected'].format(**DEFAULT_VALUES),  \
                'Expected enrollment request message does not match'
 
     return
